# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TiliGo Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Business {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  image_url?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  is_open: boolean;
  /** @nullable */
  delivery_fee?: number | null;
  /** @nullable */
  delivery_time?: string | null;
  /** @nullable */
  rating?: number | null;
  status: string;
  created_at?: string;
}

export interface BusinessInput {
  name: string;
  description?: string;
  category?: string;
  image_url?: string;
  address?: string;
  phone: string;
  password: string;
  delivery_fee?: number;
  delivery_time?: string;
}

export interface BusinessUpdate {
  name?: string;
  description?: string;
  category?: string;
  image_url?: string;
  address?: string;
  phone?: string;
  is_open?: boolean;
  delivery_fee?: number;
  delivery_time?: string;
  rating?: number;
  status?: string;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface Product {
  id: number;
  business_id: number;
  /** @nullable */
  business_name?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  image_url?: string | null;
  is_available: boolean;
  created_at?: string;
}

export interface ProductInput {
  business_id: number;
  business_name?: string;
  name: string;
  description?: string;
  price: number;
  category?: string;
  image_url?: string;
  is_available?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  image_url?: string;
  is_available?: boolean;
}

export interface Offer {
  id: number;
  business_id: number;
  /** @nullable */
  business_name?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  original_price?: number | null;
  offer_price: number;
  /** @nullable */
  image_url?: string | null;
  items_included?: string[];
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  valid_until?: string | null;
  is_active: boolean;
  created_at?: string;
}

export interface OfferInput {
  business_id: number;
  business_name?: string;
  title: string;
  description?: string;
  original_price?: number;
  offer_price: number;
  image_url?: string;
  items_included?: string[];
  badge?: string;
  valid_until?: string;
  is_active?: boolean;
}

export interface OfferUpdate {
  title?: string;
  description?: string;
  original_price?: number;
  offer_price?: number;
  image_url?: string;
  items_included?: string[];
  badge?: string;
  valid_until?: string;
  is_active?: boolean;
}

export interface OrderItem {
  /** @nullable */
  id?: number | null;
  name: string;
  price: number;
  qty: number;
}

export interface Order {
  id: number;
  order_code: string;
  business_id: number;
  business_name: string;
  customer_name: string;
  customer_phone: string;
  customer_address: string;
  /** @nullable */
  customer_lat?: number | null;
  /** @nullable */
  customer_lng?: number | null;
  /** @nullable */
  notes?: string | null;
  items?: OrderItem[];
  total: number;
  delivery_fee: number;
  status: string;
  payment_method?: string;
  priority?: boolean;
  /** @nullable */
  coupon_code?: string | null;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  delivery_id?: number | null;
  /** @nullable */
  delivery_name?: string | null;
  /** @nullable */
  delivery_lat?: number | null;
  /** @nullable */
  delivery_lng?: number | null;
  created_at?: string;
}

export interface OrderInput {
  order_code: string;
  business_id: number;
  business_name: string;
  customer_name: string;
  customer_phone: string;
  customer_address: string;
  customer_lat?: number;
  customer_lng?: number;
  notes?: string;
  items: OrderItem[];
  total: number;
  delivery_fee: number;
  payment_method?: string;
  priority?: boolean;
  coupon_code?: string;
  discount?: number;
}

export interface OrderUpdate {
  status?: string;
  delivery_id?: number;
  delivery_name?: string;
  delivery_lat?: number;
  delivery_lng?: number;
}

export interface Delivery {
  id: number;
  name: string;
  phone: string;
  vehicle: string;
  /** @nullable */
  image_url?: string | null;
  status: string;
  is_available: boolean;
  created_at?: string;
}

export interface DeliveryInput {
  name: string;
  phone: string;
  vehicle: string;
  image_url?: string;
  password: string;
}

export interface DeliveryUpdate {
  name?: string;
  phone?: string;
  vehicle?: string;
  image_url?: string;
  status?: string;
  is_available?: boolean;
}

export interface Coupon {
  id: number;
  code: string;
  /** @nullable */
  description?: string | null;
  discount_amount: number;
  uses_left: number;
  is_active: boolean;
  created_at?: string;
}

export interface CouponInput {
  code: string;
  description?: string;
  discount_amount: number;
  uses_left?: number;
  is_active?: boolean;
}

export interface CouponUpdate {
  code?: string;
  description?: string;
  discount_amount?: number;
  uses_left?: number;
  is_active?: boolean;
}

export interface AppSetting {
  id: number;
  key: string;
  value: string;
  /** @nullable */
  label?: string | null;
}

export interface AppSettingInput {
  key: string;
  value: string;
  label?: string;
}

export interface SupportTicket {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  /** @nullable */
  category?: string | null;
  priority: string;
  status: string;
  created_at?: string;
}

export interface SupportTicketInput {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
  category?: string;
  priority?: string;
}

export interface SupportTicketUpdate {
  status?: string;
  priority?: string;
}

export interface DashboardSummary {
  total_businesses: number;
  active_businesses: number;
  pending_businesses: number;
  total_couriers: number;
  approved_couriers: number;
  pending_couriers: number;
  total_orders: number;
  active_orders: number;
  delivered_orders: number;
  today_revenue: number;
  total_revenue: number;
  open_tickets: number;
}

export type ListBusinessesParams = {
status?: string;
category?: string;
};

export type ListProductsParams = {
business_id?: number;
is_available?: boolean;
};

export type ListOffersParams = {
business_id?: number;
is_active?: boolean;
};

export type ListOrdersParams = {
business_id?: number;
delivery_id?: number;
customer_phone?: string;
status?: string;
};

export type ListDeliveriesParams = {
status?: string;
is_available?: boolean;
};