# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TiliGo Delivery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all approved businesses
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Register a new business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      operationId: updateBusiness
      tags: [businesses]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    delete:
      operationId: deleteBusiness
      tags: [businesses]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /businesses/login:
    post:
      operationId: loginBusiness
      tags: [businesses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"

  # ─── PRODUCTS ──────────────────────────────────────────────────────
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      parameters:
        - in: query
          name: business_id
          schema:
            type: integer
        - in: query
          name: is_active
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    patch:
      operationId: updateOffer
      tags: [offers]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ORDERS ────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      parameters:
        - in: query
          name: business_id
          schema:
            type: integer
        - in: query
          name: delivery_id
          schema:
            type: integer
        - in: query
          name: customer_phone
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders/track/{code}:
    get:
      operationId: trackOrder
      tags: [orders]
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── DELIVERIES ────────────────────────────────────────────────────
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /coupons/validate/{code}:
    get:
      operationId: validateCoupon
      tags: [coupons]
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"

  # ─── APP SETTINGS ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Business:
      type: object
      properties:
        id:           { type: integer }
        name:         { type: string }
        description:  { type: ["string", "null"] }
        category:     { type: ["string", "null"] }
        image_url:    { type: ["string", "null"] }
        address:      { type: ["string", "null"] }
        phone:        { type: ["string", "null"] }
        is_open:      { type: boolean }
        delivery_fee: { type: ["number", "null"] }
        delivery_time: { type: ["string", "null"] }
        rating:       { type: ["number", "null"] }
        status:       { type: string }
        created_at:   { type: string }
      required: [id, name, is_open, status]
    BusinessInput:
      type: object
      properties:
        name:         { type: string }
        description:  { type: string }
        category:     { type: string }
        image_url:    { type: string }
        address:      { type: string }
        phone:        { type: string }
        password:     { type: string }
        delivery_fee: { type: number }
        delivery_time: { type: string }
      required: [name, phone, password]
    BusinessUpdate:
      type: object
      properties:
        name:         { type: string }
        description:  { type: string }
        category:     { type: string }
        image_url:    { type: string }
        address:      { type: string }
        phone:        { type: string }
        is_open:      { type: boolean }
        delivery_fee: { type: number }
        delivery_time: { type: string }
        rating:       { type: number }
        status:       { type: string }
    LoginInput:
      type: object
      properties:
        phone:    { type: string }
        password: { type: string }
      required: [phone, password]

    # ── Product ───────────────────────────────────────────────────────
...
    Offer:
      type: object
      properties:
        id:             { type: integer }
        business_id:    { type: integer }
        business_name:  { type: ["string", "null"] }
        title:          { type: string }
        description:    { type: ["string", "null"] }
        original_price: { type: ["number", "null"] }
        offer_price:    { type: number }
        image_url:      { type: ["string", "null"] }
        items_included: { type: array, items: { type: string } }
        badge:          { type: ["string", "null"] }
        valid_until:    { type: ["string", "null"] }
        is_active:      { type: boolean }
        created_at:     { type: string }
      required: [id, business_id, title, offer_price, is_active]
    OfferInput:
      type: object
      properties:
        business_id:    { type: integer }
        business_name:  { type: string }
        title:          { type: string }
        description:    { type: string }
        original_price: { type: number }
        offer_price:    { type: number }
        image_url:      { type: string }
        items_included: { type: array, items: { type: string } }
        badge:          { type: string }
        valid_until:    { type: string }
        is_active:      { type: boolean }
      required: [business_id, title, offer_price]
    OfferUpdate:
      type: object
      properties:
        title:          { type: string }
        description:    { type: string }
        original_price: { type: number }
        offer_price:    { type: number }
        image_url:      { type: string }
        items_included: { type: array, items: { type: string } }
        badge:          { type: string }
        valid_until:    { type: string }
        is_active:      { type: boolean }

    # ── Order ─────────────────────────────────────────────────────────
    OrderItem:
      type: object
      properties:
        id:    { type: ["integer", "null"] }
        name:  { type: string }
        price: { type: number }
        qty:   { type: integer }
      required: [name, price, qty]
    Order:
      type: object
      properties:
        id:               { type: integer }
        order_code:       { type: string }
        business_id:      { type: integer }
        business_name:    { type: string }
        customer_name:    { type: string }
        customer_phone:   { type: string }
        customer_address: { type: string }
        customer_lat:     { type: ["number", "null"] }
        customer_lng:     { type: ["number", "null"] }
        notes:            { type: ["string", "null"] }
        items:            { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
        total:            { type: number }
        delivery_fee:     { type: number }
        status:           { type: string }
        payment_method:   { type: string }
        priority:         { type: boolean }
        coupon_code:      { type: ["string", "null"] }
        discount:         { type: ["number", "null"] }
        delivery_id:      { type: ["integer", "null"] }
        delivery_name:    { type: ["string", "null"] }
        delivery_lat:     { type: ["number", "null"] }
        delivery_lng:     { type: ["number", "null"] }
        created_at:       { type: string }
      required: [id, order_code, business_id, business_name, customer_name, customer_phone, customer_address, total, delivery_fee, status]
    OrderInput:
      type: object
      properties:
        order_code:       { type: string }
        business_id:      { type: integer }
        business_name:    { type: string }
        customer_name:    { type: string }
        customer_phone:   { type: string }
        customer_address: { type: string }
        customer_lat:     { type: number }
        customer_lng:     { type: number }
        notes:            { type: string }
        items:            { type: array, items: { $ref: "#/components/schemas/OrderItem" } }
        total:            { type: number }
        delivery_fee:     { type: number }
        payment_method:   { type: string }
        priority:         { type: boolean }
        coupon_code:      { type: string }
        discount:         { type: number }
      required: [order_code, business_id, business_name, customer_name, customer_phone, customer_address, items, total, delivery_fee]
    OrderUpdate:
      type: object
      properties:
        status:       { type: string }
        delivery_id:  { type: integer }
        delivery_name: { type: string }
        delivery_lat: { type: number }
        delivery_lng: { type: number }

    # ── Delivery ──────────────────────────────────────────────────────
...
    Coupon:
      type: object
      properties:
        id:              { type: integer }
        code:            { type: string }
        description:     { type: ["string", "null"] }
        discount_amount: { type: number }
        uses_left:       { type: integer }
        is_active:       { type: boolean }
        created_at:      { type: string }
      required: [id, code, discount_amount, uses_left, is_active]
    CouponInput:
      type: object
      properties:
        code:            { type: string }
        description:     { type: string }
        discount_amount: { type: number }
        uses_left:       { type: integer }
        is_active:       { type: boolean }
      required: [code, discount_amount]
    CouponUpdate:
      type: object
      properties:
        code:            { type: string }
        description:     { type: string }
        discount_amount: { type: number }
        uses_left:       { type: integer }
        is_active:       { type: boolean }

    # ── AppSetting ────────────────────────────────────────────────────
...