# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TiliGo Delivery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessInput {
  name: string;
  description?: string;
  category?: string;
  image_url?: string;
  address?: string;
  phone: string;
  password: string;
  delivery_fee?: number;
  delivery_time?: string;
}